+++ src/background.js
@@ -1,56 +1,36 @@
-if (typeof browser === "undefined") {
-    let settings;
-    chrome.action.onClicked.addListener(function(tab) {
-        let url;
-        chrome.storage.local.get({settings: {home_is_prompts: true}}, function(result) {
-            settings = result.settings
-            if (settings.hasOwnProperty('home_is_prompts')) {
-                if (settings.home_is_prompts === true) {
-                    url = "pages/prompts.html"
-                }
-                else {
-                    url = "pages/explorer.html"
-                }
-            }
-            else {
-                url = "pages/explorer.html"
-            }
-            chrome.tabs.create({url: url});
-        });
-    });
-    browser = chrome;
+if (typeof browser !== "undefined") {
+    chrome.action = browser.browserAction
 }
-else {
-    let settings;
-        // Listen for a click on the browser action
-    browser.browserAction.onClicked.addListener(function(tab) {
-        browser.storage.local.get({settings: {home_is_prompts: false}}, function(result) {
-            settings = result.settings
-            let url;
-            if (settings.hasOwnProperty('home_is_prompts')) {
-                if (settings.home_is_prompts === true) {
-                    url = "pages/prompts.html"
-                }
-                else{
-                    url = "pages/explorer.html"
-                }
+let settings;
+// Listen for a click on the browser action
+chrome.action.onClicked.addListener(function(tab) {
+    chrome.storage.local.get({settings: {home_is_prompts: true}}, function(result) {
+        settings = result.settings
+        let url;
+        if (settings.hasOwnProperty('home_is_prompts')) {
+            if (settings.home_is_prompts === true) {
+                url = "pages/prompts.html"
             }
             else{
                 url = "pages/explorer.html"
             }
-            browser.tabs.create({url: url});
-        });
+        }
+        else{
+            url = "pages/prompts.html"
+        }
+        chrome.tabs.create({url: url});
     });
-}
+});
 
-browser.runtime.onMessage.addListener( async function(message, sender, sendResponse) {
+
+chrome.runtime.onMessage.addListener( async function(message, sender, sendResponse) {
     if (message.type === 'b_continue_convo') {
         console.log('background received')
-        browser.tabs.create({url: 'https://chat.openai.com/chat', active: true}, function (my_tab){
+        chrome.tabs.create({url: 'https://chat.openai.com/chat', active: true}, function (my_tab){
             let sent = false;
-            browser.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
+            chrome.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
                 if (tab.id === my_tab.id && changeInfo.status === 'complete' && !sent) {
-                    setTimeout(() => browser.tabs.sendMessage(my_tab.id, {
+                    setTimeout(() => chrome.tabs.sendMessage(my_tab.id, {
                         type: 'c_continue_convo',
                         id: message.id,
                         convo: message.convo
@@ -62,11 +42,11 @@ browser.runtime.onMessage.addListener( async function(message, sender, sendRespo
     }
 	else if(message.type ==='b_use_prompt') {
 		console.log('background received')
-        browser.tabs.create({url: 'https://chat.openai.com/chat', active: true}, function (my_tab){
+        chrome.tabs.create({url: 'https://chat.openai.com/chat', active: true}, function (my_tab){
             let sent = false;
-            browser.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
+            chrome.tabs.onUpdated.addListener(function(tabId, changeInfo, tab) {
                 if (tab.id === my_tab.id && changeInfo.status === 'complete' && !sent) {
-                    setTimeout(() => browser.tabs.sendMessage(my_tab.id, {
+                    setTimeout(() => chrome.tabs.sendMessage(my_tab.id, {
                         type: 'c_use_prompt',
                         id: message.id,
                         prompt: message.prompt
@@ -123,8 +103,8 @@ function new_prompt(title, text, tags="", category="") {
     };
     return prompt;
 }
-browser.runtime.onInstalled.addListener(async () => {
-    browser.contextMenus.create({
+chrome.runtime.onInstalled.addListener(async () => {
+    chrome.contextMenus.create({
         id: "savePrompt",
         title: "Save text as prompt",
         contexts: ["selection"],
@@ -133,14 +113,14 @@ browser.runtime.onInstalled.addListener(async () => {
 });
 
 
-browser.contextMenus.onClicked.addListener(function(info, tab) {
+chrome.contextMenus.onClicked.addListener(function(info, tab) {
     if (info.menuItemId === "savePrompt") {
-        browser.storage.local.get({prompts: []}, function(result) {
+        chrome.storage.local.get({prompts: []}, function(result) {
             let prompts = result.prompts
             prompts.push(new_prompt("", info.selectionText))
-            browser.storage.local.set({prompts: prompts})
-            browser.tabs.create({url: "pages/prompts.html"});
-            setTimeout(() => browser.runtime.sendMessage({message: "New Prompt"}), 300)
+            chrome.storage.local.set({prompts: prompts})
+            chrome.tabs.create({url: "pages/prompts.html"});
+            setTimeout(() => chrome.runtime.sendMessage({message: "New Prompt"}), 300)
         });
     }
 });